# api-gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Smart City Management System resources",
        "title": "Smart City Management System API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/transport/incidents/report": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Report a transportation-related incident.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transport"
                ],
                "summary": "Report an incident",
                "parameters": [
                    {
                        "description": "Incident Report Data",
                        "name": "incident",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transport.IncidentCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/transport.Incident"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/transport/maintenance/schedule": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve the maintenance schedule for transport infrastructure.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transport"
                ],
                "summary": "Get maintenance schedule",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/transport.MaintanceScheduleList"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/transport/parking-lots": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve information about all parking lots.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transport"
                ],
                "summary": "Get all parking lots",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/transport.PakingLotList"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/transport/parking-lots/{lotId}/reserve": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Reserve a parking lot by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transport"
                ],
                "summary": "Reserve a parking lot",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Lot ID",
                        "name": "lotId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Parking Lot Reservation Data",
                        "name": "reservation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transport.ParkingLotCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/transport.ParkingLot"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/transport/parking-lots/{lotId}/status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve status information for a specific parking lot.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transport"
                ],
                "summary": "Get parking lot status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Lot ID",
                        "name": "lotId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/transport.ParkingLot"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/transport/roads/{roadId}/signals": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create or update signals for a specific road.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transport"
                ],
                "summary": "Control road signals",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Road ID",
                        "name": "roadId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Road Signal Control Data",
                        "name": "signals",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transport.RoadCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/transport.Road"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/transport/roads/{roadId}/traffic": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve traffic information for a specific road.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transport"
                ],
                "summary": "Get road traffic status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Road ID",
                        "name": "roadId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/transport.Road"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/transport/routes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all transport routes.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transport"
                ],
                "summary": "Get all transport routes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/transport.RouteList"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/transport/routes/{routeId}/schedule": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve the schedule for a specific transport route.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transport"
                ],
                "summary": "Get route schedule",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Route ID",
                        "name": "routeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/transport.RouteScheduleList"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "transport.Incident": {
            "type": "object",
            "properties": {
                "CreatedAt": {
                    "type": "string"
                },
                "DeletedAt": {
                    "type": "string"
                },
                "Description": {
                    "type": "string"
                },
                "IncidentId": {
                    "type": "string"
                },
                "RoadId": {
                    "type": "string"
                },
                "Status": {
                    "type": "string"
                },
                "Time": {
                    "type": "string"
                },
                "UpdatedAt": {
                    "type": "string"
                }
            }
        },
        "transport.IncidentCreate": {
            "type": "object",
            "properties": {
                "Description": {
                    "type": "string"
                },
                "RoadId": {
                    "type": "string"
                },
                "Status": {
                    "type": "string"
                },
                "Time": {
                    "type": "string"
                }
            }
        },
        "transport.MaintanceSchedule": {
            "type": "object",
            "properties": {
                "CreatedAt": {
                    "type": "string"
                },
                "DeletedAt": {
                    "type": "string"
                },
                "EndPoint": {
                    "type": "string"
                },
                "EndTime": {
                    "type": "string"
                },
                "Id": {
                    "type": "string"
                },
                "Organizer": {
                    "type": "string"
                },
                "RoadId": {
                    "type": "string"
                },
                "StartPoint": {
                    "type": "string"
                },
                "StartTime": {
                    "type": "string"
                },
                "Status": {
                    "type": "string"
                },
                "TotalArea": {
                    "type": "integer"
                },
                "UpdatedAt": {
                    "type": "string"
                }
            }
        },
        "transport.MaintanceScheduleList": {
            "type": "object",
            "properties": {
                "Count": {
                    "type": "integer"
                },
                "Schedules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/transport.MaintanceSchedule"
                    }
                }
            }
        },
        "transport.PakingLotList": {
            "type": "object",
            "properties": {
                "Count": {
                    "type": "integer"
                },
                "ParkingLots": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/transport.ParkingLot"
                    }
                }
            }
        },
        "transport.ParkingLot": {
            "type": "object",
            "properties": {
                "Address": {
                    "type": "string"
                },
                "CreatedAt": {
                    "type": "string"
                },
                "DeletedAt": {
                    "type": "string"
                },
                "FreeSpaces": {
                    "type": "integer"
                },
                "LotId": {
                    "type": "string"
                },
                "LotName": {
                    "type": "string"
                },
                "OccupiedSpaces": {
                    "type": "integer"
                },
                "TotalSpaces": {
                    "type": "integer"
                },
                "UpdatedAt": {
                    "type": "string"
                }
            }
        },
        "transport.ParkingLotCreate": {
            "type": "object",
            "properties": {
                "Address": {
                    "type": "string"
                },
                "FreeSpaces": {
                    "type": "integer"
                },
                "LotName": {
                    "type": "string"
                },
                "OccupiedSpaces": {
                    "type": "integer"
                },
                "TotalSpaces": {
                    "type": "integer"
                }
            }
        },
        "transport.Road": {
            "type": "object",
            "properties": {
                "CreatedAt": {
                    "type": "string"
                },
                "DeletedAt": {
                    "type": "string"
                },
                "EndPoint": {
                    "type": "string"
                },
                "Lenght": {
                    "type": "integer"
                },
                "RoadId": {
                    "type": "string"
                },
                "RoadName": {
                    "type": "string"
                },
                "StartPoint": {
                    "type": "string"
                },
                "UpdatedAt": {
                    "type": "string"
                }
            }
        },
        "transport.RoadCreate": {
            "type": "object",
            "properties": {
                "EndPoint": {
                    "type": "string"
                },
                "Lenght": {
                    "type": "integer"
                },
                "RoadId": {
                    "type": "string"
                },
                "RoadName": {
                    "type": "string"
                },
                "StartPoint": {
                    "type": "string"
                }
            }
        },
        "transport.Route": {
            "type": "object",
            "properties": {
                "CreatedAt": {
                    "type": "string"
                },
                "DeletedAt": {
                    "type": "string"
                },
                "EndPoint": {
                    "type": "string"
                },
                "RouteId": {
                    "type": "string"
                },
                "RouteName": {
                    "type": "string"
                },
                "RouteType": {
                    "type": "string"
                },
                "StartPoint": {
                    "type": "string"
                },
                "UpdatedAt": {
                    "type": "string"
                }
            }
        },
        "transport.RouteList": {
            "type": "object",
            "properties": {
                "Routes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/transport.Route"
                    }
                }
            }
        },
        "transport.RouteSchedule": {
            "type": "object",
            "properties": {
                "ArrivalTime": {
                    "type": "string"
                },
                "CreatedAt": {
                    "type": "string"
                },
                "DayOfWeek": {
                    "type": "integer"
                },
                "DeletedAt": {
                    "type": "string"
                },
                "DepartureTime": {
                    "type": "string"
                },
                "RouteId": {
                    "type": "string"
                },
                "ScheduleId": {
                    "type": "string"
                },
                "UpdatedAt": {
                    "type": "string"
                }
            }
        },
        "transport.RouteScheduleList": {
            "type": "object",
            "properties": {
                "Schedules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/transport.RouteSchedule"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}